Rebasing onto dev/Fixing conflicts in Mercury scene
Ethan had originally created this branch and did the work to add the spaceship to the Mercury scene. For some reason that I can't remember I closed the PR, and the branch got deleted so I am restoring the branch and rebasing it onto dev to fix my mistake. The mercury scene still has Ethans work.